Move bundled zopfli sources into a new zopfli/ subdirectory.
Separating source files by origin (and license) makes it easy for users who
want to build with an external zopfli library to remove the bundled sources,
ensuring they are not used. Unbundling libraries is mandatory for packaging in
many popular Linux distributions, and can be useful for other applications as
well. Additional changes will be needed to make it easy to select an external
zopfli library.